Changes the way the tracks are stored in the DB
Since the database schema for the tracks changed, the table now contains
a record for each positional message that is to be included in the
tracks. The GeoJSON result is now generated on the API instead of using
PostGIS `ST_AsGeoJSON`.

A new format parameter is included as well, allowing the clients to
request GeoJSON lines (which are a bit more difficult to work with, but
easier to display) or
raw points (which are supper easy to process, but difficult to display
nicely).

We also imported some changes from the api-events project regarding SQL
debug logs.